move typescriptAttributeLocator file
