+++ src/typescriptAttributeLocator.ts
